+++ flutter_figma/lib/figma.dart
@@ -1,3 +1,4 @@
 export 'src/figma.dart';
-export 'src/design.dart';
+export 'src/design/design.dart';
+export 'src/design/storage.dart';
 export 'src/helpers/api_extensions.dart';