+++ figma_theme/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.1] - 2020/10/06
+
+* Added asset package.
+
 ## [0.1.0] - 2020/09/24
 
 * Initial release